+++ src/frame.rs
@@ -148,6 +148,7 @@ pub mod transport {
     pub struct PCI(u8);
 
     /// Type of the `PCI` byte
+    #[derive(Copy, Clone, Debug, PartialEq, Eq)]
     pub enum PCIType {
         /// Single Frame
         SF = 0,
@@ -162,8 +163,8 @@ pub mod transport {
     impl PCI {
         /// Create a `PCI` with type `PCIType::SF` and the given length
         pub fn new_sf(length: u8) -> PCI {
-            assert!(length <= 5, "Maximum length for single frame is 5");
-            PCI(length + 1)
+            assert!(length <= 6, "Maximum length for single frame is 6");
+            PCI(length)
         }
 
         /// Get the `PCIType` of the PCI
@@ -178,7 +179,7 @@ pub mod transport {
 
         /// Get the length field of the `PCI`
         pub const fn get_length(self) -> u8 {
-            self.0
+            self.0 & 0x0F
         }
     }
 
@@ -202,7 +203,7 @@ pub mod transport {
         // If a PDU is not completely filled the unused bytes shall be filled with 0xFF.
         let mut frame_data = [0xFFu8; 8];
         frame_data[0] = nad.0;
-        frame_data[1] = PCI::new_sf(data.len() as u8).0;
+        frame_data[1] = PCI::new_sf(data.len() as u8 + 1).0;
         frame_data[2] = sid.0;
         frame_data[3..data.len() + 3].clone_from_slice(data);
         Frame::from_data(pid, &frame_data)
@@ -341,6 +342,7 @@ pub mod diagnostic {
 #[cfg(test)]
 mod tests {
     use super::diagnostic::*;
+    use super::transport::*;
     use super::*;
 
     struct CheckSumTestData<'a> {
@@ -428,6 +430,13 @@ mod tests {
     #[should_panic]
     fn test_pid_from_id_panic() {
         PID::from_id(64);
+    }
+
+    #[test]
+    fn test_pci() {
+        let pci = PCI::new_sf(5);
+        assert_eq!(pci.get_type(), PCIType::SF);
+        assert_eq!(pci.get_length(), 5);
     }
 
     #[test]